backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
pymupdf
python-docx
sentence-transformers
faiss-cpu
numpy
pydantic
